HapPea/requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.8
django-allauth==0.39.1
django-heroku==0.3.1
django-phonenumber-field==2.3.1
django-sslserver==0.20
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
phonenumbers==8.10.9
Pillow==6.0.0
psycopg2==2.8.1
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
whitenoise==4.1.2
